LayoutTests/inspector/styles/styles-iframe-expected.txt

CONSOLE MESSAGE: line 9: iframe loaded
Tests that proper (and different) styles are returned for body elements of main document and iframe.


Main frame style:
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
media="screen" (styles-iframe.html)
#main  { (styles-iframe.html:5)
background: blue;

[expanded] 
div  { (user agent stylesheet)
display: block;


iframe style:
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
media="screen" (styles-iframe-data.html)
body  { (styles-iframe-data.html:4)
background: green;

[expanded] 
body  { (user agent stylesheet)
display: block;
margin: 8px;
